% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-funs.R
\name{bc_correct}
\alias{bc_correct}
\title{Error correct BC sequences with known list}
\usage{
bc_correct(inbcs, bc_list, maxdist = 1, cores = 1)
}
\arguments{
\item{inbcs}{input character vector of uncorrected BC sequences}

\item{bc_list}{input character vector of expected known BC sequences}

\item{maxdist}{edit distance (hamming) used for collapse, Default: 1}

\item{cores}{Number of CPU cores to use. Default: 1}
}
\value{
returns a vector of error-corrected sequences with same length as input sequences.
}
\description{
bc_correct performs error correction on an input vector of raw sequences and a list of expected sequences.
}
\details{
NA
}
\examples{
\dontrun{
bc_correct(in_strings, known_strings, maxdist = 1)
}
}
\seealso{
\code{\link[data.table]{data.table-package}}
}
