% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-funs.R
\name{extract_complex_spike_dat}
\alias{extract_complex_spike_dat}
\title{Load and extract spUMIs from sequencing reads with complex molecular spikes set (5')}
\usage{
extract_complex_spike_dat(
  bam_path,
  bc_df,
  spike_groundtruth = NULL,
  max_pattern_dist = 1,
  cores = 12,
  min_mapq_value = 255,
  fixed_start_pos = NULL
)
}
\arguments{
\item{bam_path}{path to input bam file, must be indexed zUMIs output file}

\item{bc_df}{data.frame containing expected spike-in names & barcodes}

\item{spike_groundtruth}{data.table containing all known spike-in molecules (NOT IMPLEMENTED yet) Default: NULL}

\item{max_pattern_dist}{number of sequencing errors allowed in the sequence pattern recognition used to extract barcode & spUMIs. Default:1}

\item{cores}{number of CPU cores used. Default: 12}

\item{min_mapq_value}{minimum MAPQ mapping quality (default only uniquely aligned reads). Default: 255}

\item{fixed_start_pos}{require fixed starting position of BC/spUMI sequence in the read (given as integer). Default: NULL}
}
\value{
returns a data.table with reads, their UMI and the raw & error-corrected spUMI for each spike-in sequence and barcode.
}
\description{
extract_complex_spike_dat is used to extract and parse molecular spike reads from bam files.
}
\details{
Barcodes are error corrected allowing 1 hamming distance.
}
\examples{
\dontrun{
example_dat <- extract_complex_spike_dat(
 bam_path = bam,
 bc_df = spike_info,
 max_pattern_dist = 2,
 cores = 11
)
}
}
\seealso{
\code{\link[Rsamtools]{BamInput}},\code{\link[Rsamtools]{ScanBamParam}}
\code{\link[data.table]{data.table-package}}
}
