% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-funs.R
\name{get_overrepresented_spikes}
\alias{get_overrepresented_spikes}
\alias{plot_spike_distances}
\title{get_overrepresented_spikes}
\usage{
get_overrepresented_spikes(dat, readcutoff = 100, nbccutoff = 5)

plot_spike_distances(dat, threads = 32)
}
\arguments{
\item{dat}{}
}
\value{
blubb

blubb
}
\description{
bla

bla
}
\details{
XYZ

XYZ
}
\examples{
\dontrun{
get_overrepresented_spikes(...
}
\dontrun{
get_overrepresented_spikes(...
}
}
\seealso{
\code{\link[data.table]{data.table-package}}

\code{\link[data.table]{data.table-package}}
}
