% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-funs.R
\name{plot_spike_distances}
\alias{plot_spike_distances}
\title{Plot minimal pairwise distances between spUMIs within / over cells}
\usage{
plot_spike_distances(dat, threads = 32)
}
\arguments{
\item{dat}{input data.table of loaded spUMI data. Must have BC and spikeUMI columns.}

\item{threads}{number of threads to use for edit distance calculation, Default: 32}
}
\value{
returns a ggplot object
}
\description{
Plot minimal pairwise distances between spUMIs within and over cells to determine the appropriate error-correction edit distance for spUMI sequences
}
\details{
uses hamming distance as the edit distance measure.
}
\examples{
\dontrun{
plot_spike_distances(spikedat)
}
}
\seealso{
\code{\link[data.table]{data.table-package}}
}
