% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-funs.R
\name{subsample_recompute}
\alias{subsample_recompute}
\title{Subsample molecular spikes per cell and recompute hamming distance correction}
\usage{
subsample_recompute(dat, mu_nSpikeUMI, threads = 8)
}
\arguments{
\item{dat}{input data.table of loaded spUMI data. Must have BC,UX and spikeUMI_hd2 columns.}

\item{mu_nSpikeUMI}{mean expression level of the spUMI to subsample to.}

\item{threads}{number of threads to use, Default: 8}
}
\value{
returns ...
}
\description{
return_corrected_umi performs error correction on an input vector of raw UMI sequences.
}
\details{
For each cell barcode, subsample the number of spUMIs to an expression level of the given mean and recomputes the error corrected UMIs.
}
\examples{
\dontrun{
subsample_recompute(spikedat, mu_nSpikeUMI)
}
}
\seealso{
\code{\link[data.table]{data.table-package}}
}
