% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-package.R
\docType{package}
\name{UMIcountR-package}
\alias{UMIcountR}
\alias{UMIcountR-package}
\title{UMIcountR: tools for spike-in UMI analysis}
\description{
Provide tools for extraction, filtering and analysis of sequencing data containing molecular spike-ins.
}
\author{
\strong{Maintainer}: Christoph Ziegenhain \email{christoph.ziegenhain@ki.se} (\href{https://orcid.org/0000-0003-2208-4877}{ORCID})

}
\keyword{internal}
