% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMIcountR-funs.R
\name{return_corrected_umi}
\alias{return_corrected_umi}
\title{Return corrected UMI sequences}
\usage{
return_corrected_umi(
  umi_input,
  editham = 1,
  collapse_mode = NULL,
  ngram_split = NULL
)
}
\arguments{
\item{umi_input}{input character vector of uncorrected UMI sequences}

\item{editham}{edit distance (hamming) used for collapse, Default: 1}

\item{collapse_mode}{collapse mode to use, Default: adjacency}
}
\value{
returns a vector of error-corrected UMIs with same length as input sequences. Implemented collapse algorithms: "adjacency","adjacency_directional","adjacency_singleton","cluster"
}
\description{
return_corrected_umi performs error correction on an input vector of raw UMI sequences.
}
\details{
For a description of implemented collapse algorithms, please refer to Ziegenhain, Hendriks et al., 2021.
}
\examples{
\dontrun{
return_corrected_umi(UX_strings, editham = 1, collapse_mode = "adjacency")
}
}
\seealso{
\code{\link[data.table]{data.table-package}}
}
